celestia: move ExtendedHeader parser to specific_helpers, fatal on empty chain id
- ParseCelestiaExtendedHeader / CelestiaExtendedHeader live in specific_helpers (shared by head polling, chain validator, lower-bound detector), tests moved along
- chain validator: any mismatch incl. an empty chain id is FatalSettingError, no separate branch
- comment on GetFinalizedBlock: CometBFT commits are final, the head is the finalized pointer
